+++ src/lua/CommunityBalanceMod/BalanceValueChanges/Balance.lua
@@ -5,6 +5,8 @@ kPulseGrenadeDamage = 60 -- vanilla: 50
 kPulseGrenadeEnergyDamageRadius = 6 -- 4
 kDropMineCost = 5 --7
 kWelderDropCost = 2 -- 7
+kStabEnergyCost = 25 --30
+kStabResearchCost = 20 -- 25
 
 kAdvancedMarineSupportResearchCost = 15 -- 20
 kNanoShieldCost = 2 --3
@@ -13,6 +15,7 @@ kNanoShieldCost = 2 --3
 -- Nerfs
 kClusterGrenadeDamageRadius = 8 --10
 kClusterFragmentDamageRadius = 5 -- 6
+